confirm delete recordings

+++ app/assets/javascripts/controllers/mixins/recordable_list.coffee
@@ -4,4 +4,10 @@ App.RecordableList = Ember.Mixin.create
   ).property('connector.isConnected')
 
   destroy: (recording) ->
-    recording.deleteRecord()
+    if confirm('Are you sure you want to delete this recording')
+      recording.deleteRecord()
+
+      # we need to call remove object here because the contents might be
+      # created using a findQuery and deleteRecord will only remove the
+      # object from the findAll record array
+      @get('model').removeObject(recording)
